Add draft to repositores
We want to follow the clear architecture principles and split some
concepts and abstraction between presentation, business logic and data
layers. The domain module represents the business logic layer and must
have the entities and the spec to how to access data and process it into
the entities to be drawn into the screens. The repository pattern is our
choice to abstract the data source.