Add usecases to domain module
This is the very first step to represent usecase in our project. The
idea behind it is represent each 'action' in the app as usecase. This
approach help us to test our code because each use case has a single
action. Once it's a action, we can reuse a given use case several times
when we need. The usecase will be responsabel to some business rule or
retrieve/send data from repository.