Add data-local module
The data-local module will be resposable to persist senstive user
information in a encrypted place. We are aware that is not a good
practice. We should keep only the session token to retrieve a new access
token but it is a requirement of the business team and it's ok for test
purpose.

We also created a token repository to persist the access token when the
app is killed by the OS.