Improve GetToken usecase
Now we have a robust use case and it allow us to retrieve data from
repository and observe it at view or somewhere else. The GetToken
expects username and password as a parameter and will returns a
observable that emits a Token. We will use the 'execute' fun to handle
with the observable flow.