Use GetToken to make the login
Now we are 'connecting the dots'. Now we have a screen that uses a
usecase to retrieve login infomartion and the use case has a concrete
instance of the repository that calls the login endpoint and retrieve
token info. It also fire a exception if the credential does not match.

We will improve this approach soon, using viewmodels to remove such
logic from view.